% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{plot.simutraj}
\alias{plot.simutraj}
\title{Plot an object of class \code{simutraj}.}
\usage{
\method{plot}{simutraj}(x, ...)
}
\arguments{
\item{x}{An object of \code{simutraj} class resulting from running a simulator of trajectories built using the \code{build_simulator} function.}
\item{\dots}{Further arguments to be passed to \code{plot} or to \code{plot.simutraj}.}
}
\value{
\code{plot.simutraj} returns a graphical plot of the trajectory, 
i.e the number if individuals of each compartment through time.
}
\description{
Plot an object of class \code{simutraj}.
}
\author{
Gonche Danesh
}
