\name{TIPS-package}
\alias{TIPS-package}
\alias{TIPS}
\docType{package}
\title{
\packageTitle{TIPS}
}
\description{
\packageDescription{TIPS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TIPS}
\packageIndices{TIPS}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{TIPS}

Maintainer: \packageMaintainer{TIPS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
