\name{TPIS-package}
\alias{TPIS-package}
\alias{TPIS}
\docType{package}
\title{
\packageTitle{TPIS}
}
\description{
\packageDescription{TPIS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TPIS}
\packageIndices{TPIS}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{TPIS}

Maintainer: \packageMaintainer{TPIS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
