% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{plot.simutraj}
\alias{plot.simutraj}
\title{Plot an object of class \code{simutraj}.}
\usage{
\method{plot}{simutraj}(x, ...)
}
\arguments{
\item{x}{An object of \code{simutraj} resulting from running a simulator of trajectories built using the \code{build_simulator} function.}

\item{...}{Arguments to be passed to methods, such as graphical parameters}
}
\description{
Plot an object of class \code{simutraj}.
}
\author{
Gonche Danesh
}
